remove nav bars, remove extra buttons, move connect to device tab
