+++ EnlightenMobile/Views/DeviceView.xaml
@@ -3,7 +3,7 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml" 
              xmlns:vm="clr-namespace:EnlightenMobile.ViewModels"
              x:Class="EnlightenMobile.Views.DeviceView"
-             Title="{Binding title}"
+             Shell.NavBarIsVisible="False"
              BackgroundColor="#fff">
 
     <ContentPage.BindingContext>
@@ -18,6 +18,7 @@
 
             <!-- BLE Device Info -->
             <Label FontSize="Large" Text="BLE Device Info"/>
+            <Button Text="{Binding bleBtnTxt}" Margin="5" Clicked="connectPage"/>
             <StackLayout Margin="10, 0, 10, 0">
                 <FlexLayout JustifyContent="SpaceBetween">
                     <Label Text="Device Name"/>
