+++ EnlightenMobile/Views/SettingsView.xaml
@@ -7,8 +7,7 @@
              xmlns:telerikPrimitives="clr-namespace:Telerik.XamarinForms.Primitives;assembly=Telerik.XamarinForms.Primitives"
              xmlns:telerikChart="clr-namespace:Telerik.XamarinForms.Chart;assembly=Telerik.XamarinForms.Chart"
              xmlns:telerikInput="clr-namespace:Telerik.XamarinForms.Input;assembly=Telerik.XamarinForms.Input"
-
-             Title="{Binding title}"
+             Shell.NavBarIsVisible="False"
              BackgroundColor="White">
              <!-- Title="💾" -->
